+++ app/client/keybase/debug/keystore.go
@@ -36,11 +36,14 @@ func init() {
 
 func initializeDebugKeybase() error {
 	// Create/Open the keybase at `$HOME/.pocket/keys`
-	kb, err := keybase.NewKeybase(debugKeybasePath)
+	kb, err := keybase.NewBadgerKeybase(debugKeybasePath)
 	if err != nil {
 		return err
 	}
-	db := kb.GetBadgerDB()
+	db, err := kb.GetBadgerDB()
+	if err != nil {
+		return err
+	}
 
 	if err := restoreBadgerDB(build.DebugKeybaseBackup, db); err != nil {
 		return err